jit nqueens

+++ rust/src/syms.rs
@@ -1,8 +1,8 @@
 #[derive(Default, Clone)]
 pub struct Symbol<'a> {
     name: &'a str,
-    addr: u64,
-    size: u64
+    addr: i64,
+    size: i64
 }
 
 pub type Symbols<'a> = Vec<Symbol<'a>>;
@@ -19,8 +19,8 @@ pub fn get_symbols<'a>(elf_file: &elf::ElfBytes<'a, elf::endian::AnyEndian>) -> 
         if let Ok(name) = strtab.get(sym.st_name as usize) {
             symbols.push(Symbol {
                 name,
-                addr: sym.st_value,
-                size: sym.st_size
+                addr: sym.st_value as i64,
+                size: sym.st_size as i64
             });
         }
     }
@@ -31,8 +31,8 @@ pub fn get_symbols<'a>(elf_file: &elf::ElfBytes<'a, elf::endian::AnyEndian>) -> 
 
 #[derive(Debug, Clone)]
 pub struct SymbolTreeNode<'a> {
-    start: u64,
-    size: u64,
+    start: i64,
+    size: i64,
     name: &'a str,
     left: Option<Box<SymbolTreeNode<'a>>>,
     right: Option<Box<SymbolTreeNode<'a>>>
@@ -55,7 +55,7 @@ impl<'a> SymbolTreeNode<'a> {
         }))
     }
 
-    pub fn lookup(&self, value: u64) -> Option<(&'a str, u64)> {
+    pub fn lookup(&self, value: i64) -> Option<(&'a str, i64)> {
         if value < self.start {
             return self.left.as_ref().and_then(|node| node.lookup(value))
         }
